move .cpp to source/ root directory, refactor JsonAux to a namespace instead of a class and more networking stuff
