+++ EZVIDIA/include/utils/StringUtils.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include<string>
+
+namespace StringUtils {
+	std::wstring stringToWideString(std::string str);
+	std::string wideStringToString(std::wstring wstr);
+}
